+++ src/Ops.Service/Filters/RequestFilter.cs
@@ -1,0 +1,11 @@
+﻿namespace Ocuda.Ops.Service.Filters
+{
+    public class RequestFilter : BaseFilter
+    {
+        public RequestFilter(int page) : base(page)
+        {
+        }
+
+        public bool? IsProcessed { get; set; }
+    }
+}