+++ src/PolarisHelper/PolarisContext.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace Ocuda.PolarisHelper
+{
+    public class PolarisContext : DbContext
+    {
+        public PolarisContext(DbContextOptions options) : base(options) { }
+    }
+}